00 main de mentira

+++ FitQuestPrototype/src/main/java/com/salesianos/FitQuestPrototype/Entrenamiento/Dto/Ejercicio/GetEjercicioDto.java
@@ -6,8 +6,8 @@ public record GetEjercicioDto(
         Long id,
         String nombre,
         String descripcion,
-        double series,
-        double repeticiones,
+        int series,
+        int repeticiones,
         double duracion,
         String urlImagen
 ) {
